requirements.txt
# This code runs on python 3.10
geopandas==0.12.2
shapely==2.0.1
pystac-client==0.6.0
odc-stac==0.3.5
xarray==2023.1.0
rioxarray==0.13.4
numpy==1.24.4
pandas==1.5.3
zarr==2.14.2
gdal==3.6.2
rasterio==1.3.4
osmnx ==1.9.1
boto3 ==1.34.40